requirements.txt
anthropic>=0.40
python-dotenv>=1.0
trafilatura>=2.0
requests>=2.31
